requirements.txt
click
pycrypto
